+++ Changelog.md
@@ -18,6 +18,7 @@ YogaSMC Changelog
 - NC: hide selected text which status is revealed by image
 - YogaVPC: notify timestamp for key press
 - NC: fix easter egg and enable it by default
+- YogaSMC: support atomicSpDeciKelvinKey
 
 #### v1.3.0
 - Pane: fix loading prior to 10.14.4 with embedded swift runtime, thx @Charlyo